+++ examples/Backend/ChargePaymentController.php
@@ -1,45 +1,46 @@
 <?php
-/**
- * This is the controller for the Card recurring example.
- * It is called when the pay button on the index page is clicked.
+
+/*
+ *  Controller for charge on payment (capture).
  *
- * Copyright (C) 2020 - today Unzer E-Com GmbH
+ *  Copyright (C) 2022 - today Unzer E-Com GmbH
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *  http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
  *
- * @link  https://docs.unzer.com/
+ *  @link  https://docs.unzer.com/
  *
- * @package  UnzerSDK\examples
+ *  @package  UnzerSDK\examples
+ *
  */
 
 /** Require the constants of this example */
-require_once __DIR__ . '/Constants.php';
+require_once __DIR__ . '/../Constants.php';
 
 /** Require the composer autoloader file */
 /** @noinspection PhpIncludeInspection */
 require_once __DIR__ . '/../../../../autoload.php';
 
-use UnzerSDK\Constants\RecurrenceTypes;
 use UnzerSDK\examples\ExampleDebugHandler;
 use UnzerSDK\Exceptions\UnzerApiException;
+use UnzerSDK\Resources\CustomerFactory;
 use UnzerSDK\Resources\TransactionTypes\Charge;
 use UnzerSDK\Unzer;
 
 session_start();
-session_unset();
 
 $clientMessage = 'Something went wrong. Please try again later.';
 $merchantMessage = 'Something went wrong. Please try again later.';
+$debugHandler = new ExampleDebugHandler();
 
 function redirect($url, $merchantMessage = '', $clientMessage = '')
 {
@@ -50,37 +51,30 @@ function redirect($url, $merchantMessage = '', $clientMessage = '')
 }
 
 // You will need the id of the payment type created in the frontend (index.php)
-if (!isset($_POST['resourceId'])) {
+if (!isset($_POST['payment_id'])) {
     redirect(FAILURE_URL, 'Resource id is missing!', $clientMessage);
 }
-$paymentTypeId   = $_POST['resourceId'];
-
-// Just for this example: Use selected recurrence type. Scheduled will be used as default.
-$recurrenceType = $_POST['recurrence_type'] ?? RecurrenceTypes::SCHEDULED;
+$paymentId   = $_POST['payment_id'];
 
 // Catch API errors, write the message to your log and show the ClientMessage to the client.
 try {
     // Create an Unzer object using your private key and register a debug handler if you want to.
     $unzer = new Unzer(UNZER_PAPI_PRIVATE_KEY);
-    $unzer->setDebugMode(true)->setDebugHandler(new ExampleDebugHandler());
-    $paymentType = $unzer->fetchPaymentType($paymentTypeId);
+    $unzer->setDebugMode(true)->setDebugHandler($debugHandler);
 
-    $charge = (new Charge(12.99, 'EUR', MY_RETURN_CONTROLLER_URL))
-        ->setRecurrenceType($recurrenceType, $paymentType);
-
-    $transaction = $unzer->performCharge($charge, $paymentTypeId);
+    $transaction = $unzer->performChargeOnPayment($paymentId, new Charge());
 
     // You'll need to remember the paymentId for later in the ReturnController (in case of 3ds)
-    $_SESSION['PaymentTypeId'] = $paymentTypeId;
     $_SESSION['ShortId'] = $transaction->getShortId();
-    $_SESSION['recurrenceType'] = $transaction->getRecurrenceType();
+    unset($_SESSION['isAuthorizeTransaction']);
 
-    // Redirect to the 3ds page or to success depending on the state of the transaction
+    // Redirect to the failure page or to success depending on the state of the transaction
     $redirect = !empty($transaction->getRedirectUrl());
     if (!$redirect && $transaction->isSuccess()) {
-        redirect(SUCCESS_URL);
+        redirect(BACKEND_URL);
     } elseif ($redirect && $transaction->isPending()) {
-        redirect($transaction->getRedirectUrl());
+        redirect(BACKEND_FAILURE_URL, 'Transaction initiated by merchant should not redirect to 3ds Page. The customer needs to
+        do the 3ds authentication first for that payment type.');
     }
 
     // Check the result message of the transaction to find out what went wrong.
@@ -91,4 +85,4 @@ try {
 } catch (RuntimeException $e) {
     $merchantMessage = $e->getMessage();
 }
-redirect(FAILURE_URL, $merchantMessage, $clientMessage);
+redirect(BACKEND_FAILURE_URL, $merchantMessage, $clientMessage);