+++ examples/Bancontact/index.php
@@ -51,7 +51,9 @@ require_once __DIR__ . '/../../../../autoload.php';
         <!-- The Holder field UI Element will be inserted here -->
     </div>
     <div class="field" id="error-holder" style="color: #9f3a38"> </div>
-    <button class="unzerUI primary button fluid" id="submit-button" type="submit">Pay</button>
+    <div class="field">
+        <button class="unzerUI primary button fluid" id="submit-button" type="submit">Pay</button>
+    </div>
 </form>
 
 <script>