+++ examples/Klarna/Controller.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * This is the controller for the Hosted Payment Page example.
+ * This is the controller for the Klarna example.
  * It is called when the pay button on the index page is clicked.
  *
- * Copyright (C) 2020 - today Unzer E-Com GmbH
+ * Copyright (C) 2022 - today Unzer E-Com GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,14 +30,16 @@ require_once __DIR__ . '/Constants.php';
 require_once __DIR__ . '/../../../../autoload.php';
 
 use UnzerSDK\Constants\Salutations;
+use UnzerSDK\Constants\ShippingTypes;
 use UnzerSDK\examples\ExampleDebugHandler;
 use UnzerSDK\Exceptions\UnzerApiException;
-use UnzerSDK\Resources\EmbeddedResources\Address;
-use UnzerSDK\Unzer;
 use UnzerSDK\Resources\Basket;
 use UnzerSDK\Resources\CustomerFactory;
+use UnzerSDK\Resources\EmbeddedResources\Address;
 use UnzerSDK\Resources\EmbeddedResources\BasketItem;
-use UnzerSDK\Resources\PaymentTypes\Paypage;
+use UnzerSDK\Resources\TransactionTypes\Authorization;
+use UnzerSDK\Resources\TransactionTypes\Charge;
+use UnzerSDK\Unzer;
 
 session_start();
 session_unset();
@@ -53,7 +55,12 @@ function redirect($url, $merchantMessage = '', $clientMessage = '')
     die();
 }
 
-// These lines are just for this example
+// You will need the id of the payment type created in the frontend (index.php)
+if (!isset($_POST['resourceId'])) {
+    redirect(FAILURE_URL, 'Resource id is missing!', $clientMessage);
+}
+$paymentTypeId   = $_POST['resourceId'];
+
 $transactionType = $_POST['transaction_type'] ?? 'authorize';
 
 // Catch API errors, write the message to your log and show the ClientMessage to the client.
@@ -62,44 +69,9 @@ try {
     $unzer = new Unzer(UNZER_PAPI_PRIVATE_KEY);
     $unzer->setDebugMode(true)->setDebugHandler(new ExampleDebugHandler());
 
-    // A customer with matching addresses is mandatory for Installment payment type
-    $address  = (new Address())
-        ->setName('Max Mustermann')
-        ->setStreet('Vangerowstr. 18')
-        ->setCity('Heidelberg')
-        ->setZip('69155')
-        ->setCountry('DE');
-
-    // Create a charge/authorize transaction
-    $customer = CustomerFactory::createCustomer('Max', 'Mustermann')
-        ->setSalutation(Salutations::MR)
-        ->setBirthDate('2000-02-12')
-        ->setLanguage('de')
-        ->setEmail('test@test.com');
-
-    // These are the mandatory parameters for the payment page ...
-    $paypage = new Paypage(119.00, 'EUR', RETURN_CONTROLLER_URL);
-
     $orderId = 'o' . str_replace(['0.', ' '], '', microtime(false));
 
-    // ... however you can customize the Payment Page using additional parameters.
-    $paypage->setShopName('My Test Shop')
-        ->setShopDescription('Best shop in the whole world!')
-        ->setTagline('Try and stop us from being awesome!')
-        ->setOrderId('OrderNr' . $orderId)
-        ->setTermsAndConditionUrl('https://www.unzer.com/en/')
-        ->setPrivacyPolicyUrl('https://www.unzer.com/de/datenschutz/')
-        ->setImprintUrl('https://www.unzer.com/de/impressum')
-        ->setHelpUrl('https://www.unzer.com/de/support')
-        ->setContactUrl('https://www.unzer.com/en/ueber-unzer')
-        ->setFullPageImage(UNZER_PP_FULL_PAGE_IMAGE_URL)
-        ->setLogoImage(UNZER_PP_LOGO_URL)
-        ->setInvoiceId('i' . microtime(true));
-
-    // ... in order to enable Unzer Instalment you will need to set the effectiveInterestRate as well.
-    $paypage->setEffectiveInterestRate(4.99);
-
-    // ... a Basket is mandatory for InstallmentSecured
+    // A Basket is mandatory for Invoice Secured payment type
     $basketItem = (new BasketItem())
         ->setAmountPerUnitGross(119.00)
         ->setVat(19.00)
@@ -112,22 +84,49 @@ try {
         ->addBasketItem($basketItem)
         ->setCurrencyCode('EUR');
 
-    if ($transactionType === 'charge') {
-        $unzer->initPayPageCharge($paypage, $customer, $basket);
-    } else {
-        // For demo purpose we set customer address for authorize only to enable instalment payment.
-        // That way e.g. Invoice secured will display customer form on payment page.
-        $customer
-            ->setShippingAddress($address)
-            ->setBillingAddress($address);
-        $unzer->initPayPageAuthorize($paypage, $customer, $basket);
+    // A customer is mandatory for Klarna payment type
+    $address  = (new Address())
+        ->setName('Max Universum')
+        ->setStreet('Hugo-Junkers-Str. 4')
+        ->setZip('60386')
+        ->setCity('Frankfurt am Main')
+        ->setCountry('DE')
+        ->setState('DE-BO')
+        ->setShippingType(ShippingTypes::EQUALS_BILLING);
+
+    $customer = CustomerFactory::createCustomer('Peter', 'Universum')
+        ->setSalutation(Salutations::MR)
+        ->setCompany('Unzer GmbH')
+        ->setBirthDate('1989-12-24')
+        ->setEmail('peter.universum@universum-group.de')
+        ->setMobile('+49172123456')
+        ->setPhone('+4962216471100')
+        ->setBillingAddress($address)
+        ->setLanguage('de');
+
+    $transactionData = [119.00, 'EUR', RETURN_CONTROLLER_URL];
+    $additionalTransactionData = [
+        'termsAndConditionUrl' => 'https://www.unzer.com/de/',
+        'privacyPolicyUrl' => 'https://www.unzer.com/de/'
+    ];
+
+    // Create a authorize transaction to get the redirectUrl.
+    $authorizationInstance = (new Authorization(...$transactionData))
+        ->setAdditionalTransactionData((object)$additionalTransactionData);
+    $transaction = $unzer->performAuthorization($authorizationInstance, $paymentTypeId, $customer, null, $basket);
+
+    // You'll need to remember the paymentId for later in the ReturnController
+    $_SESSION['PaymentId'] = $transaction->getPaymentId();
+    $_SESSION['ShortId']   = $transaction->getShortId();
+    $_SESSION['isAuthorizeTransaction'] = true;
+
+    // Redirect to the Klarna page
+    if (!$transaction->isError() && $transaction->getRedirectUrl() !== null) {
+        redirect($transaction->getRedirectUrl());
     }
 
-    $_SESSION['PaymentId'] = $paypage->getPaymentId();
-
-    // Redirect to the paypage
-    redirect($paypage->getRedirectUrl());
-
+    // Check the result message of the charge to find out what went wrong.
+    $merchantMessage = $transaction->getMessage()->getCustomer();
 } catch (UnzerApiException $e) {
     $merchantMessage = $e->getMerchantMessage();
     $clientMessage = $e->getClientMessage();