+++ examples/Wechatpay/index.php
@@ -53,7 +53,9 @@ require_once __DIR__ . '/../../../../autoload.php';
 
 <form id="payment-form" class="unzerUI form" novalidate>
     <div class="field" id="error-holder" style="color: #9f3a38"> </div>
-    <button class="unzerUI primary button fluid" id="submit-button" type="submit">Pay</button>
+    <div class="field">
+        <button class="unzerUI primary button fluid" id="submit-button" type="submit">Pay</button>
+    </div>
 </form>
 
 <script>