+++ test/BaseIntegrationTest.php
@@ -63,7 +63,7 @@ class BaseIntegrationTest extends BasePaymentTest
      *
      * @throws \UnzerSDK\Exceptions\UnzerApiException
      */
-    protected function createPaylaterAuthorization(): Authorization
+    protected function createPaylaterInvoiceAuthorization(): Authorization
     {
         $paylaterInvoice = $this->unzer->createPaymentType(new PaylaterInvoice());
 
@@ -75,5 +75,13 @@ class BaseIntegrationTest extends BasePaymentTest
 
         $authorization = $this->unzer->performAuthorization($authorization, $paylaterInvoice, $customer, null, $basket);
         return $authorization;
+    }
+
+    /**
+     * @return void
+     */
+    protected function useNon3dsKey(): void
+    {
+        $this->getUnzerObject()->setKey(EnvironmentService::getTestPrivateKey(true));
     }
 }